python/requirements.txt
glfw>=2.6.2
PyOpenGL>=3.1.6
numpy>=1.26.0
noise>=1.2.2
